+++ src/Constants.lua
@@ -7,6 +7,9 @@ local addonName, Yipper = ...
 
 Yipper.Constants = {}
 
+-- Track our version in case we need to do something breaking
+Yipper.Constants.VERSION = "1.4.1"
+
 -- Define the list of chat events we want to listen to
 Yipper.Constants.ChatEvents = {
     "CHAT_MSG_EMOTE",